Courses save without clicking generate schedules
* Use throttling to limit how many save_courses requests are sent
