+++ autoscheduler/frontend/src/components/SchedulingPage/ConfigureCard/ConfigureCard.tsx
@@ -13,7 +13,7 @@ import Meeting from '../../../types/Meeting';
 import { parseAllMeetings } from '../../../redux/actions/courseCards';
 // DEBUG
 import { RootState } from '../../../redux/reducer';
-import { CourseCardArray, CustomizationLevel, SerializedCourseCardOptions } from '../../../types/CourseCardOptions';
+import { CourseCardArray, CustomizationLevel } from '../../../types/CourseCardOptions';
 import Availability from '../../../types/Availability';
 import { formatTime } from '../../../utils/timeUtil';
 
@@ -103,29 +103,6 @@ const ConfigureCard: React.FC = () => {
         dispatch(selectSchedule(0));
         if (ref.current) setLoading(false);
       });
-
-    // make request to save course cards
-    const courseData: SerializedCourseCardOptions[] = [];
-    for (let i = 0; i < courseCards.numCardsCreated; i++) {
-      const course = courseCards[i];
-      if (course) {
-        const sections = course.sections.filter(({ selected }) => selected).map((sectionSel) => (
-          sectionSel.section.id
-        ));
-        courseData.push({
-          course: course.course,
-          customizationLevel: course.customizationLevel,
-          honors: course.honors,
-          web: course.web,
-          sections,
-        });
-      }
-    }
-
-    fetch('sessions/save_courses', {
-      method: 'PUT',
-      body: JSON.stringify(courseData),
-    });
   }, [avsList, courseCards, dispatch, includeFull, term]);
 
   return (
