+++ autoscheduler/frontend/src/components/SchedulingPage/CourseSelectColumn/CourseSelectColumn.tsx
@@ -6,6 +6,7 @@ import { RootState } from '../../../redux/reducer';
 import { CourseCardArray, SerializedCourseCardOptions } from '../../../types/CourseCardOptions';
 import CourseSelectCard from './CourseSelectCard/CourseSelectCard';
 import { addCourseCard, replaceCourseCards, clearCourseCards } from '../../../redux/actions/courseCards';
+import throttle from '../../../utils/throttle';
 
 /**
  * Renders a column of CourseSelectCards, as well as a button to add course cards
@@ -19,8 +20,8 @@ const CourseSelectColumn: React.FC = () => {
 
   const term = useSelector<RootState, string>((state) => state.term);
 
+  // When term is changed, fetch saved courses for the new term
   React.useEffect(() => {
-    // When term is changed, fetch saved courses for the new term
     if (term) {
       fetch(`sessions/get_saved_courses?term=${term}`).then((res) => (
         res.json()
@@ -32,6 +33,44 @@ const CourseSelectColumn: React.FC = () => {
     // on unmount, clear course cards
     return (): void => { dispatch(clearCourseCards()); };
   }, [term, dispatch]);
+
+  // When courseCards is changed, save courses (and throttle API calls)
+  React.useEffect(() => {
+    if (!term) return;
+
+    const saveCourses = (): void => {
+      // Serialize courseCards and make API call
+      const courses: SerializedCourseCardOptions[] = [];
+      for (let i = 0; i < courseCards.numCardsCreated; i++) {
+        const course = courseCards[i];
+        if (course) {
+          const sections = course.sections.filter(({ selected }) => selected).map((sectionSel) => (
+            sectionSel.section.id
+          ));
+          courses.push({
+            course: course.course,
+            customizationLevel: course.customizationLevel,
+            honors: course.honors,
+            web: course.web,
+            sections,
+          });
+        }
+      }
+
+      const body = JSON.stringify({
+        courses,
+        term,
+      });
+
+      fetch('sessions/save_courses', {
+        method: 'PUT',
+        body,
+      });
+    };
+
+    throttle(`saveCourseCards${term}`, saveCourses, 15000, true);
+  }, [courseCards, term]);
+
 
   const rows: JSX.Element[] = [];
